+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add phpstan.neon file
+- add cpsell project words for Gitlab-CI
+
+### Changed
+- remove usage of deprecated ContainerAware class
+- add getRequestStack & getTitleResolver to TwigExtensionBase
+- improve type-hinting using RevisionableStorageInterface for Loader::loadEntityRevision && Render::renderEntityRevision
+- remove usage of deprecated twig_date_converter function when Twig 3.9 is installed
+
+### Fixed
+- fix tests running on 10.3-dev with rendered webp image
 
 ## [6.0.1] - 2024-03-01
 ### Changed